I realized we can use uint16 as here instead

+++ src/cpu.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 /* memory specific */
-uint8 m_peek8(CPU *u, int n)
+uint8 m_peek8(CPU *u, uint16 n)
 {
     return u->mem.content[n];
 }
@@ -14,7 +14,7 @@ uint8 m_consume8(CPU *u)
     return u->mem.content[u->mem.ptr++];
 }
 
-uint16 m_peek16(CPU *u, int n)
+uint16 m_peek16(CPU *u, uint16 n)
 {
     /* https://stackoverflow.com/a/1935457 */
     return m_peek8(u, n) + (m_peek8(u, n + 1) << 8);
